neurolight/requirements.txt
PySide6>=6.6.0
opencv-python>=4.8.0
Pillow>=10.0.0
ultralytics>=8.0.0
numpy>=1.24.0
scipy>=1.11.0
matplotlib>=3.8.0
statsmodels>=0.14.0
imageio>=2.31.0

